+++ src/components/index.js
@@ -3,7 +3,23 @@ import Sidebar from './Sidebar';
 import Footer from './Footer';
 import ProjectModal from './ProjectModal';
 import BugsModal from './BugsModal';
-import Hero from './Hero';
 import FormAlert from './FormAlert';
+import PageTitle from './PageTitle';
+import ProjectCard from './ProjectCard';
+import ProjectUtility from './ProjectUtility';
+import BugUtility from './BugUtility';
+import BugCard from './BugCard';
 
-export { Navbar, Sidebar, Footer, ProjectModal, BugsModal, Hero, FormAlert };
+export {
+  Navbar,
+  Sidebar,
+  Footer,
+  ProjectModal,
+  BugsModal,
+  FormAlert,
+  PageTitle,
+  ProjectCard,
+  BugCard,
+  ProjectUtility,
+  BugUtility,
+};
